Move type definitions and functions related to attachment and viewer to ViewerModule.
